pyproject.toml
[tool.poetry]
name = "jwtrestapiflask"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.2.0"
python-dotenv = "^0.21.0"
flask = "^2.2.2"
pyjwt = "^2.6.0"
redis = "^4.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
